// src/types/component-types-sb.d.ts
// This file was generated by the storyblok CLI.
// DO NOT MODIFY THIS FILE BY HAND.
import type { ISbStoryData } from 'storyblok';
export interface FeatureStoryblok {
  name?: string;
  component: 'feature';
  _uid: string;
  [k: string]: any;
}

export interface GridStoryblok {
  columns?: (
    | FeatureStoryblok
    | GridStoryblok
    | GroupStoryblok
    | NavItemStoryblok
    | PageStoryblok
    | TeaserStoryblok
  )[];
  component: 'grid';
  _uid: string;
  [k: string]: any;
}

export type MultiassetStoryblok = {
  alt: string | null;
  copyright?: string | null;
  fieldtype: 'asset';
  id: number;
  filename: string | null;
  name: string;
  title: string | null;
  focus: string | null;
  meta_data?: {
    [k: string]: any;
  };
  source?: string | null;
  is_external_url?: boolean;
  is_private?: boolean;
  src?: string;
  updated_at?: string;
  width?: number | null;
  height?: number | null;
  aspect_ratio?: number | null;
  public_id?: string | null;
  content_type?: string;
  [k: string]: any;
}[];

export interface GroupStoryblok {
  name?: string;
  items?: (
    | FeatureStoryblok
    | GridStoryblok
    | GroupStoryblok
    | NavItemStoryblok
    | PageStoryblok
    | TeaserStoryblok
  )[];
  images?: MultiassetStoryblok;
  component: 'Group';
  _uid: string;
  [k: string]: any;
}

export interface AssetStoryblok {
  alt: string | null;
  copyright?: string | null;
  fieldtype: 'asset';
  id: number;
  filename: string | null;
  name: string;
  title: string | null;
  focus: string | null;
  meta_data?: {
    [k: string]: any;
  };
  source?: string | null;
  is_external_url?: boolean;
  is_private?: boolean;
  src?: string;
  updated_at?: string;
  width?: number | null;
  height?: number | null;
  aspect_ratio?: number | null;
  public_id?: string | null;
  content_type?: string;
  [k: string]: any;
}

export interface NavItemStoryblok {
  label?: string;
  description?: string;
  dropdown?: (
    | FeatureStoryblok
    | GridStoryblok
    | GroupStoryblok
    | NavItemStoryblok
    | PageStoryblok
    | TeaserStoryblok
  )[];
  link?: string;
  image?: AssetStoryblok;
  component: 'navItem';
  _uid: string;
  [k: string]: any;
}

export interface PageStoryblok {
  body?: (
    | FeatureStoryblok
    | GridStoryblok
    | GroupStoryblok
    | NavItemStoryblok
    | PageStoryblok
    | TeaserStoryblok
  )[];
  component: 'page';
  _uid: string;
  [k: string]: any;
}

export interface TeaserStoryblok {
  headline?: string;
  component: 'teaser';
  _uid: string;
  [k: string]: any;
}